add reshuffle check in order and tasks

+++ auto_bot/handlers/order/handlers.py
@@ -16,7 +16,7 @@ from auto_bot.handlers.order.keyboards import inline_spot_keyboard, inline_route
     inline_add_info_kb, inline_change_currency_trip, personal_order_start_kb, personal_order_time_kb, \
     personal_order_end_kb, personal_order_back_kb
 from auto_bot.handlers.order.utils import buttons_addresses, text_to_client, validate_text, get_geocoding_address, \
-    save_location_to_redis
+    save_location_to_redis, check_reshuffle
 from auto_bot.main import bot
 from scripts.conversion import get_address, get_location_from_db
 from auto_bot.handlers.order.static_text import *
@@ -433,45 +433,44 @@ def handle_callback_order(update, context):
     query = update.callback_query
     data = query.data.split(' ')
     driver = Driver.get_by_chat_id(chat_id=query.from_user.id)
+    vehicle = check_reshuffle(driver)
     order = Order.objects.filter(pk=int(data[1])).first()
     if order.status_order in (Order.COMPLETED, Order.IN_PROGRESS):
         query.edit_message_text(text=already_accepted)
         return
     if order.status_order == Order.ON_TIME:
-        if driver:
-            if driver.vehicle:
-                order.driver = driver
-                order.save()
-                group_msg = redis_instance().hget('group_msg', order.pk)
-                context.bot.delete_message(chat_id=ParkSettings.get_value('ORDER_CHAT'),
-                                           message_id=group_msg)
-                redis_instance().hdel('group_msg', order.pk)
-                for manager in Manager.objects.exclude(chat_id__isnull=True):
-                    if redis_instance().hexists(str(manager.chat_id), f'personal {order.id}'):
-                        context.bot.send_message(chat_id=manager.chat_id,
-                                                 text=f"Замовлення {order.id} прийнято")
-                        redis_instance().hdel(str(manager.chat_id), f'personal {order.id}')
-                report_for_client = client_order_text(driver, driver.vehicle.name, driver.vehicle.licence_plate,
-                                                      driver.phone_number, order.sum)
-                message_info = redis_instance().hget(str(order.chat_id_client), 'client_msg')
-                client_msg = text_to_client(order, report_for_client, delete_id=message_info,
-                                            button=inline_reject_order(order.pk))
-                redis_instance().hset(str(order.chat_id_client), 'client_msg', client_msg)
-                driver_msg = context.bot.send_message(
-                    chat_id=driver.chat_id,
-                    text=time_order_accepted(order.from_address, timezone.localtime(order.order_time).time()))
-                redis_instance().hset(str(driver.chat_id), 'driver_msg', driver_msg.message_id)
-            else:
-                context.bot.send_message(chat_id=query.from_user.id, text=add_many_auto_text)
+        if driver and vehicle:
+            order.driver = driver
+            order.save()
+            group_msg = redis_instance().hget('group_msg', order.pk)
+            context.bot.delete_message(chat_id=ParkSettings.get_value('ORDER_CHAT'),
+                                       message_id=group_msg)
+            redis_instance().hdel('group_msg', order.pk)
+            for manager in Manager.objects.exclude(chat_id__isnull=True):
+                if redis_instance().hexists(str(manager.chat_id), f'personal {order.id}'):
+                    context.bot.send_message(chat_id=manager.chat_id,
+                                             text=f"Замовлення {order.id} прийнято")
+                    redis_instance().hdel(str(manager.chat_id), f'personal {order.id}')
+            report_for_client = client_order_text(driver, vehicle.name, vehicle.licence_plate,
+                                                  driver.phone_number, order.sum)
+            message_info = redis_instance().hget(str(order.chat_id_client), 'client_msg')
+            client_msg = text_to_client(order, report_for_client, delete_id=message_info,
+                                        button=inline_reject_order(order.pk))
+            redis_instance().hset(str(order.chat_id_client), 'client_msg', client_msg)
+            driver_msg = context.bot.send_message(
+                chat_id=driver.chat_id,
+                text=time_order_accepted(order.from_address, timezone.localtime(order.order_time).time()))
+            redis_instance().hset(str(driver.chat_id), 'driver_msg', driver_msg.message_id)
         else:
-            context.bot.send_message(chat_id=query.from_user.id, text=accept_order_error)
+            context.bot.send_message(chat_id=query.from_user.id, text=add_many_auto_text)
+
     else:
         order.driver = driver
         order.save()
         markup = inline_spot_keyboard(order.latitude, order.longitude, pk=order.id)
         query.edit_message_text(text=order_info(order))
         query.edit_message_reply_markup(reply_markup=markup)
-        report_for_client = client_order_text(driver, driver.vehicle.name, driver.vehicle.licence_plate,
+        report_for_client = client_order_text(driver, vehicle.name, vehicle.licence_plate,
                                               driver.phone_number, order.sum)
         client_msg = redis_instance().hget(str(order.chat_id_client), 'client_msg')
         text_to_client(order, report_for_client, button=inline_reject_order(order.pk), message_id=client_msg)
@@ -479,10 +478,10 @@ def handle_callback_order(update, context):
         order.status_order, order.accepted_time = Order.IN_PROGRESS, timezone.localtime()
         order.save()
         if order.chat_id_client:
-            lat, long = get_location_from_db(driver.vehicle.licence_plate)
+            lat, long = get_location_from_db(vehicle.licence_plate)
             bot.send_message(chat_id=order.chat_id_client, text=order_customer_text)
             message = bot.sendLocation(order.chat_id_client, latitude=lat, longitude=long, live_period=1800)
-            send_map_to_client.delay(order.id, driver.vehicle.licence_plate, message.message_id, message.chat_id)
+            send_map_to_client.delay(order.id, vehicle.licence_plate, message.message_id, message.chat_id)
 
 
 def payment_request(update, context, chat_id_client, provider_token, url,
@@ -568,7 +567,8 @@ def handle_order(update, context):
             query.edit_message_text(text=calc_price_text)
             start_route = redis_instance().hget(str(chat_id), 'start_route')
             s, e = int(start_route), int(timezone.localtime().timestamp())
-            get_distance_trip.delay(data[1], query.message.message_id, s, e, driver.vehicle.gps_id)
+            vehicle = check_reshuffle(driver)[0]
+            get_distance_trip.delay(data[1], query.message.message_id, s, e, vehicle.gps_id)
         else:
             if order.payment_method == price_inline_buttons[4].split()[1]:
                 cash_order(update, query, order)
